requirements.txt
flask>=3.1.0
tifffile>=2024.12.12
pandas>=2.2.3
numpy>=2.2.1
matplotlib>=3.10.0
pillow>=11.1.0
scipy>=1.15.0
requests>=2.32.3
flask-session>=0.8.0
seaborn>=0.13.2